Welcome!

KadanoffBaym.jl is the first fully adaptive solver for Kadanoff-Baym equations written in Julia.

Tip

To learn more about the solver and Kadanoff-Baym equations, have a look into our accompanying paper.`

Installation

To install, use Julia's built-in package manager

julia> ] add KadanoffBaym

The most recent version of KadanoffBaym.jl requires Julia v1.7 or later.

Examples

To learn how to work with KadanoffBaym.jl, there are two options:

The examples folder of our repository, which contains notebooks for all of the systems studied in our paper.

The examples section of this documentation. If you are interested in quantum dynamics, we recommend you start with the tight-binding model. More advanced users can jump directly to Fermi-Hubbard model part I about the second Born approximation. Part II shows how to solve the more involved $T$-matrix approximation.

Note

KadanoffBaym.jl can also be used to simulate stochastic processes. An introduction to this topic is given here.

Library

KadanoffBaym.jl was designed to be lean and simple and hence only exports a handful of functions, namely GreenFunction (together with two possible time symmetries, Symmetrical and SkewHermitian) and the integrator kbsolve!.

Note

You need to import an FFT library – e.g., FFTW – to use wigner_transform.

Index

KadanoffBaym.GreenFunction

KadanoffBaym.wigner_transform

Solver

Missing docstring.

Missing docstring for kbsolve!(fv!, fd!, u0::Vector{<:GreenFunction}, (t0, tmax)). Check Documenter's build log for details.

Green Functions

KadanoffBaym.GreenFunction — Type

GreenFunction(g::AbstractArray, s::AbstractSymmetry)

A container interface for g with array indexing respecting some symmetry rule s. Because of that, g must be square in its last 2 dimensions, which can be resized with resize!.

The array g is not restricted to being contiguous. For example, g can have Matrix{T}, Array{T,4}, Matrix{SparseMatrixCSC{T}}, etc as its type.

Notes

The GreenFunction does not own g. Proper care must be taken when using multiple GreenFunctions since using the same array will result in unexpected behaviour

julia> data = zeros(2,2)
julia> g1 = GreenFunction(data, Symmetrical)
julia> g2 = GreenFunction(data, Symmetrical)
julia> g1[1,1] = 3
julia> @show g2[1,1]
julia> g1.data === g2.data # they share the same data

Indexing with less indices than the dimension of g results in a "take-all-to-the-left" indexing

julia> gf[i,j] == gf[:,:,...,:,i,j]
julia> gf[i,j,k] == gf[:,:,...,:,i,j,k]

Custom symmetries can be implemented via multiple dispatch

julia> struct MySymmetry <: KadanoffBaym.AbstractSymmetry end
julia> @inline KadanoffBaym.symmetry(::Type{MySymmetry}) = conj

Examples

GreenFunction simply takes some data g and embeds the symmetry s in its indexing

julia> time_dim = 3
julia> spin_dim = 2
julia> data = zeros(spin_dim, spin_dim, time_dim, time_dim)
julia> gf = GreenFunction(data, Symmetrical)
julia> gf[2,1] = rand(spin_dim, spin_dim)
julia> @show gf[1,2]
julia> @show KadanoffBaym.symmetry(Symmetrical)(gf[2,1])

Wigner Transformation

KadanoffBaym.wigner_transform — Method

wigner_transform(x::AbstractMatrix; ts=1:size(x,1), fourier=true)

Wigner-Ville transformation

$x_W(ω, T) = i ∫dt x(T + t/2, T - t/2) e^{+i ω t}$

or

$x_W(τ, T) = x(T + t/2, T - t/2)$

of a 2-point function x. Returns a tuple of x_W and the corresponding axes (ω, T) or (τ, T), depending on the fourier keyword.

The motivation for the Wigner transformation is that, given an autocorrelation function x, it reduces to the spectral density function at all times T for stationary processes, yet it is fully equivalent to the non-stationary autocorrelation function. Therefore, the Wigner (distribution) function tells us, roughly, how the spectral density changes in time.

Optional keyword parameters

ts::AbstractVector: Time grid for x. Defaults to a UnitRange.

fourier::Bool: Whether to Fourier transform. Defaults to true.

Notes

The algorithm only works when ts – and consequently x – is equidistant.

References

https://en.wikipedia.org/wiki/Wigner_distribution_function

http://tftb.nongnu.org

Citation

If you use KadanoffBaym.jl in your research, please cite our paper:

@Article{10.21468/SciPostPhysCore.5.2.030,
	title={{Adaptive Numerical Solution of Kadanoff-Baym Equations}},
	author={Francisco Meirinhos and Michael Kajan and Johann Kroha and Tim Bode},
	journal={SciPost Phys. Core},
	volume={5},
	issue={2},
	pages={30},
	year={2022},
	publisher={SciPost},
	doi={10.21468/SciPostPhysCore.5.2.030},
	url={https://scipost.org/10.21468/SciPostPhysCore.5.2.030},
}